pyproject.toml
[tool.poetry]
name = "snapshots"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
psycopg2-binary = "^2.9.9"
tabulate = "^0.9.0"
watchdog = "^4.0.0"
prometheus-client = "^0.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.3.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
